+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Changelog
 
 November 12, 2014
++ Fixed animated class
+
+November 12, 2014
 + Actually a plugin
 
 November 10, 2014